refactor: metadata resolver into smaller and more manageable resolvers
